refactor(shared): move reactive custom widgets to a new directory structure
