+++ lib/core/models/custom_widgets/paint_editor_widgets.dart
@@ -3,8 +3,8 @@ import 'package:flutter/widgets.dart';
 
 // Project imports:
 import '/features/paint_editor/paint_editor.dart';
-import '/shared/widgets/custom_widgets/reactive_custom_appbar.dart';
-import '/shared/widgets/custom_widgets/reactive_custom_widget.dart';
+import '../../../shared/widgets/reactive_widgets/reactive_custom_appbar.dart';
+import '../../../shared/widgets/reactive_widgets/reactive_custom_widget.dart';
 import 'utils/custom_widgets_standalone_editor.dart';
 import 'utils/custom_widgets_typedef.dart';
