+++ lib/pro_image_editor.dart
@@ -46,8 +46,8 @@ export 'shared/services/content_recorder/utils/generate_high_quality_image.dart'
 export 'shared/utils/converters.dart';
 export 'shared/utils/decode_image.dart';
 export 'shared/widgets/color_picker/bar_color_picker.dart';
-export 'shared/widgets/custom_widgets/reactive_custom_appbar.dart';
-export 'shared/widgets/custom_widgets/reactive_custom_widget.dart';
+export 'shared/widgets/reactive_widgets/reactive_custom_appbar.dart';
+export 'shared/widgets/reactive_widgets/reactive_custom_widget.dart';
 export 'shared/widgets/extended/extended_pop_scope.dart';
 export 'core/constants/editor_style_constants.dart';
