+++ lib/shared/services/import_export/import_state_history.dart
@@ -8,9 +8,9 @@ import 'package:flutter/services.dart';
 import '/core/models/editor_image.dart';
 import '/core/models/history/state_history.dart';
 import '/core/models/layers/layer.dart';
+import '/features/crop_rotate_editor/models/transform_factors.dart';
 import '/features/filter_editor/utils/filter_generator/filter_addons.dart';
-import '../../../features/crop_rotate_editor/models/transform_factors.dart';
-import '../../../features/tune_editor/models/tune_adjustment_matrix.dart';
+import '/features/tune_editor/models/tune_adjustment_matrix.dart';
 import '../../utils/parser/double_parser.dart';
 import '../../utils/parser/int_parser.dart';
 import '../../utils/parser/size_parser.dart';
