+++ lib/shared/services/layer_transform_generator.dart
@@ -2,11 +2,11 @@
 import 'dart:math';
 import 'dart:ui';
 
+import '/core/models/layers/layer.dart';
 import '/features/crop_rotate_editor/enums/crop_rotate_angle_side.dart';
+import '/features/crop_rotate_editor/models/transform_factors.dart';
 import '/features/crop_rotate_editor/utils/rotate_angle.dart';
 import '/features/main_editor/services/layer_copy_manager.dart';
-import '../../core/models/layers/layer.dart';
-import '../../features/crop_rotate_editor/models/transform_factors.dart';
 
 /// A class for generating transformed layers in an image editor.
 ///
