+++ lib/shared/widgets/reactive_widgets/reactive_custom_appbar.dart
